requirements.txt
feedparser==6.0.12
google-genai==0.3.0
newspaper3k==0.2.8
requests==2.31.0
nltk==3.8.1
pytz==2024.1
Flask==3.0.3
gunicorn==22.0.0
